add binder for write, and rename Json.class to JsonHolder.
